Backend/Document Parser/requirements.txt
fastapi
uvicorn
python-multipart
opendataloader-pdf
python-docx
pandas
requests
python-dotenv
pypdf
